added @jit for conv

+++ tests/test_mnist.py
@@ -93,7 +93,7 @@ class TestMNIST(unittest.TestCase):
         t.set_description(f"loss: {float(loss[0]):.2f} accuracy: {float(accuracy):.2f}")
 
     def numpy_eval():
-      Y_test_preds_out = model.forward(Tensor(X_test.reshape((-1, 28*28))))
+      Y_test_preds_out = model.forward(Tensor(X_test.reshape((-1, 28*28)).astype(np.float32)))
       Y_test_preds = np.argmax(Y_test_preds_out.data, axis=1)
       return (Y_test == Y_test_preds).mean()
